+++ jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
 	testEnvironment: "node",
+	testRunner: "jest-jasmine2", // https://github.com/jestjs/jest/issues/11698
 	roots: ["<rootDir>/src", "<rootDir>/test"],
 	testRegex: "(\\.|/)test\\.tsx?$",
 	moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],